move setups to test_toolkit
