--- tests/setups/.gitignore
@@ -1,1 +1,0 @@
-tests_setup.py
