--- tests/setups/virtual_drive.py
@@ -1,3 +1,0 @@
-from tests.setups.specifiers import VirtualDriveSpecifier
-
-TESTS_SETUP = VirtualDriveSpecifier(ip="127.0.0.1", port=1061)
