+++ tests/test_examples.py
@@ -27,7 +27,7 @@ from ingeniamotion.information import Information
 from ingeniamotion.motion import Motion
 from ingeniamotion.pdo import PDONetworkManager, PDOPoller
 from tests.conftest import connect_canopen, connect_ethernet, connect_soem
-from tests.setups.descriptors import (
+from tests.tests_toolkit.setups.descriptors import (
     DriveCanOpenSetup,
     DriveEcatSetup,
     DriveEthernetSetup,
