+++ tests/tests_toolkit/pytest_addoptions.py
@@ -2,7 +2,7 @@ def pytest_addoption(parser):
     parser.addoption(
         "--setup",
         action="store",
-        default="tests.setups.tests_setup.TESTS_SETUP",
+        default="tests.tests_setup.TESTS_SETUP",
         help="Module and location from which to import the setup."
         "It will default to a file that you can create on"
         "tests_setup.py inside of the folder setups with a variable called TESTS_SETUP"
