+++ tests/tests_toolkit/setups/descriptors.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
 
-from tests.setups.rack_service_client import RackServiceClient
-from tests.setups.specifiers import (
+from tests.tests_toolkit.setups.rack_service_client import RackServiceClient
+from tests.tests_toolkit.setups.specifiers import (
     Interface,
     MultiDriveConfigSpecifier,
     MultiRackServiceConfigSpecifier,
