+++ tests/tests_toolkit/setups/virtual_drive.py
@@ -1,0 +1,3 @@
+from tests.tests_toolkit.setups.specifiers import VirtualDriveSpecifier
+
+TESTS_SETUP = VirtualDriveSpecifier(ip="127.0.0.1", port=1061)
